move secret-utils to generators/base/support/secret
